Tests/LibWeb/Text/expected/Worker/Worker-requestAnimationFrame-supported.txt

nested in dedicated worker: raf callback ran
nested in shared worker: raf threw: NotSupportedError, caf threw: NotSupportedError
